index

+++ app/Http/Controllers/TasksController.php
@@ -9,18 +9,23 @@ use App\Models\Task;    // 追加
 
 class TasksController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
-        // メッセージ一覧を取得
-        $tasks = Task::all();         // 追加
-
-        // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [     // 追加
-            'tasks' => $tasks,        // 追加
-        ]);
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザーを取得
+            $user = \Auth::user();
+            // ユーザーの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザーの投稿も取得するように変更しますが、現時点ではこのユーザーの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        
+        // dashboardビューでそれらを表示
+        return view('dashboard', $data);
     }
 
     /**
@@ -28,12 +33,15 @@ class TasksController extends Controller
      */
     public function create()
     {
-        $task = new Task;
+ 
+            $task = new Task;
+    
+            // メッセージ作成ビューを表示
+            return view('tasks.create', [
+                'task' => $task,
+            ]);
 
-        // メッセージ作成ビューを表示
-        return view('tasks.create', [
-            'task' => $task,
-        ]);
+
     }
 
     /**
@@ -47,12 +55,14 @@ class TasksController extends Controller
             'content' => 'required',
         ]);
 
-        
+
         // メッセージを作成
         $task = new Task;
         $task->status = $request->status;    // 追加
         $task->content = $request->content;
+        $task->user_id = $request->user()->id;
         $task->save();
+
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -114,8 +124,13 @@ class TasksController extends Controller
     {
         // idの値でメッセージを検索して取得
         $task = task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        
+        if (\Auth::id() === $task->user_id) {
+           // メッセージを削除
+            $task->delete();
+            return back()
+             ->with('success', 'Delete Successful');
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
